+++ client/game/Empire/Planets/Planet.vue
@@ -85,6 +85,7 @@
 
             text-indent: -1000em;
 
+            // these need to be synced with /server/config/index.js
             &--gas { background-position: 0 -48px; }
             &--ice { background-position: 0 -96px; }
             &--iron { background-position: 0 -144px; }