+++ client/game/handlers/format.js
@@ -6,37 +6,28 @@
  **********************************************************************************************************************/
 
 /*
- * get Roman Number from latin number
- * I'm really lazy and we won't have more than 20 planets.
- * @param {Number} latin - [0..20]
+ * get roman string from latin number
+ * https://www.selftaughtjs.com/algorithm-sundays-converting-roman-numerals/
+ * @param {Number} latin
  * @returns {String} roman
  */
 const latinToRoman = latin => {
-    let romanNumbers = [
-        "",
-        "I",
-        "II",
-        "III",
-        "IV",
-        "V",
-        "VI",
-        "VII",
-        "VIII",
-        "IX",
-        "X",
-        "XI",
-        "XII",
-        "XIII",
-        "XIV",
-        "XV",
-        "XVI",
-        "XVII",
-        "XVIII",
-        "XIX",
-        "XX"
-    ];
-    latin = parseInt(latin, 10); // make sure its an integer
-    return romanNumbers[latin];
+    let result = "";
+    let decimal = [1000, 900, 500, 400, 100, 90, 50, 40, 10, 9, 5, 4, 1];
+    let roman = ["M", "CM", "D", "CD", "C", "XC", "L", "XL", "X", "IX", "V", "IV", "I"];
+    latin = parseInt(latin, 10);
+    if (latin === 0) return "N";
+    for (let i = 0; i <= decimal.length; i++) {
+        // looping over every element of our arrays
+        while (latin % decimal[i] < latin) {
+            // keep trying the same number until it won"t fit anymore
+            result += roman[i];
+            // add the matching roman number to our result string
+            latin -= decimal[i];
+            // remove the decimal value of the roman number from our number
+        }
+    }
+    return result;
 };
 
-export { latinToRoman };
+export {latinToRoman};